+++ @plotly/dash-component-plugins/src/index.js
@@ -1,0 +1,3 @@
+import { asyncDecorator, isReady } from './dynamicImport';
+
+export { asyncDecorator, isReady };